+++ src/main/java/io/github/poshjosh/ratelimiter/web/spring/RateLimitingFilter.java
@@ -1,10 +1,7 @@
 package io.github.poshjosh.ratelimiter.web.spring;
 
 import io.github.poshjosh.ratelimiter.RateLimiter;
-import io.github.poshjosh.ratelimiter.web.core.RateLimiterConfigurer;
-import io.github.poshjosh.ratelimiter.web.core.WebRateLimiterContext;
-import io.github.poshjosh.ratelimiter.web.core.WebRateLimiterRegistries;
-import io.github.poshjosh.ratelimiter.web.core.WebRateLimiterRegistry;
+import io.github.poshjosh.ratelimiter.web.core.*;
 import io.github.poshjosh.ratelimiter.util.RateLimitProperties;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -90,11 +87,11 @@ public abstract class RateLimitingFilter extends GenericFilterBean {
     }
 
     protected boolean tryConsume(HttpServletRequest httpRequest) {
-        return webRateLimiterRegistry.tryAcquire(httpRequest, 1);
+        return webRateLimiterRegistry.tryAcquire(RequestInfos.of(httpRequest), 1);
     }
 
-    public RateLimiter getRateLimiter(HttpServletRequest request) {
-        return webRateLimiterRegistry.getRateLimiterOrUnlimited(request);
+    public RateLimiter getRateLimiter(HttpServletRequest httpRequest) {
+        return webRateLimiterRegistry.getRateLimiterOrUnlimited(RequestInfos.of(httpRequest));
     }
 
     public WebRateLimiterRegistry getRateLimiterRegistry() {