+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Direct rate limiting using method: `RateLimiterRegistry#tryAcquire`.
 - Method `RateLimiterRegistry#isWithinLimit(K key)`.
 
+### Changed
+
+- Decoupled from `javax.servlet` by replaced `javax.servlet.HttpServletRequest` with `RequestInfo`
+
 ## [ [0.8.0](https://github.com/poshjosh/rate-limiter-spring/tree/v0.8.0) ] - 2024-08-05
 
 ### Added